Remove the :main declaration from project.clj
It doesn't make sense for libraries to have a :main namespace set, and removing
this declaration eliminates the potentially confusing warning message "Warning:
The Main-Class specified does not exist within the jar. It may not be
executable as expected. A gen-class directive may be missing in the namespace
which contains the main method" generated by 'lein uberjar' when deploying to
Heroku.