Add ClojureScript, Reagent, Figwheel
Use ClojureScript and Reagent to render the splash page, and call
through to said ClojureScript code from puzzles.web.server. Use Hiccup
to render the base HTML in puzzles.web.server.

The server can now be started in any of the following ways:
* lein do clean, figwheel
* lein do clean, cljsbuild once, repl --> (puzzles.web.server/start)
* lein do clean, cljsbuild once, run -m puzzles.web.server
* lein do clean, uberjar && heroku local

Figwheel will host a server on port 3449, while any of the other options
will use port 5000. Figwheel is the best option for development, while
using Heroku is the best way to verify that deployment will go smoothly.