Fix Emacs integration with Figwheel
Previously, an odd behavior was being observed where if Figwheel was
started from within a REPL or within Emacs, it could exist to an
existing page but when the page was reloaded, it would get a 404 error
for main.js.

Adding the :http-server-root key to the :figwheel config in project.clj
fixes this issue.