Actually fix flaky Figwheel behavior
Previous changes to downgrade figwheel-sidecar and to add
:http-server-root did not fix the problem and turned out to be
completely unrelated. It turns out that Figwheel, when launched via
Figwheel Sidecar (but not when launched via lein-figwheel), does not
serve files from resources/ unless the directory exists *before*
starting Figwheel.

Checking resources/ into version control and removing it from
:clean-targets should fix this flaky behavior. An ideal solution would
be to make Figwheel work even if resources/ does not exist when it is
launched, but in the meantime this fix will ensure correct behavior.

Since the previous fixes were misguided and introduced unneeded
complexity, they are reverted in this commmit.